DesktopApp/requirements.txt
matplotlib
PyQt5
pyserial
